Remove flowMonitor key after the MultipartReply is received (#64)
After "monitorEnabled" is enabled, an OpenFlow FlowDesc message is sent to dump
the flow stats if a user calls OFSwitch.DumpFlowStats or
Flow.MonitorRealizeStatus. At this time the message's transactionsID is added
into a concurrent map monitoredFlows. The issue is the transactionID is not
removed after the reply of FlowDesc message is received.  This change is to fix
the issue.

Signed-off-by: wenyingd <wenyingd@vmware.com>